requirements.txt
numpy
scipy
matplotlib
shapely
pyyaml
scikit-learn
imageio
pandas
